move chainIdsToAffiliates to apiState, pass cumulativeAffiliateFeeBps to route
