+++ packages/client/src/api/getChains.ts
@@ -1,4 +1,4 @@
-import { ClientState } from "../state";
+import { ClientState } from "../state/clientState";
 import { api } from "../utils/generateApi";
 
 export const chains = api({
