+++ packages/client/src/api/postRoute.ts
@@ -1,9 +1,17 @@
 import { api, ApiRequest } from "../utils/generateApi";
+import { ApiState } from "src/state/apiState";
 
-export const route = api({
-  methodName: "getRouteV2",
-  method: "post",
-  path: "/v2/fungible/route",
-});
+export const route = async (request: RouteRequest) => {
+  const requestWithAffiliateFeeBps = {
+    ...request,
+    cumulativeAffiliateFeeBps: ApiState.cumulativeAffiliateFeeBPS,
+  };
+
+  return api({
+    methodName: "getRouteV2",
+    method: "post",
+    path: "/v2/fungible/route",
+  })(requestWithAffiliateFeeBps);
+};
 
 export type RouteRequest = ApiRequest<"getRouteV2">;
