+++ packages/client/src/private-functions/cosmos/signCosmosMessageDirect.ts
@@ -6,7 +6,7 @@ import { makeAuthInfoBytes, makeSignDoc } from "@cosmjs/proto-signing/build/sign
 import { makePubkeyAnyFromAccount } from "src/proto-signing/pubkey";
 import { fromBase64 } from "@cosmjs/encoding/build/base64";
 import { Int53 } from "@cosmjs/math/build/integers";
-import { ClientState } from "src/state";
+import { ClientState } from "src/state/clientState";
 import { getEncodeObjectFromCosmosMessage } from "./getEncodeObjectFromCosmosMessage";
 import { SignCosmosMessageDirectOptions } from "src/types/client-types";
