+++ packages/client/src/private-functions/executeTransactions.ts
@@ -1,6 +1,6 @@
 import { TxResult } from "src/types/client-types";
 import { ExecuteRouteOptions } from "../public-functions/executeRoute";
-import { ClientState } from "../state";
+import { ClientState } from "../state/clientState";
 import { TransferStatus, Tx } from "../types/swaggerTypes";
 import { executeCosmosTransaction } from "./cosmos/executeCosmosTransaction";
 import { executeEvmTransaction } from "./evm/executeEvmTransaction";
