+++ packages/client/src/private-functions/getRpcEndpointForChain.ts
@@ -1,5 +1,5 @@
 import { chains, findFirstWorkingEndpoint } from "src/chains";
-import { ClientState } from "../state";
+import { ClientState } from "../state/clientState";
 
 export const getRpcEndpointForChain = async (chainId: string) => {
   if (ClientState.endpointOptions.getRpcEndpointForChain) {
