+++ packages/client/src/public-functions/getSigningStargateClient.ts
@@ -1,5 +1,5 @@
 import { OfflineSigner } from "@cosmjs/proto-signing/build/signer";
-import { ClientState } from "../state";
+import { ClientState } from "../state/clientState";
 import { SigningStargateClient } from "@cosmjs/stargate/build/signingstargateclient";
 import { accountParser } from "src/registry";
 import { getRpcEndpointForChain } from "../private-functions/getRpcEndpointForChain";
