+++ packages/client/src/public-functions/setClientOptions.ts
@@ -10,13 +10,12 @@ import { Registry } from "@cosmjs/proto-signing/build/registry";
 import { MsgExecuteContract } from "cosmjs-types/cosmwasm/wasm/v1/tx";
 import { MsgExecute } from "src/codegen/initia/move/v1/tx";
 import { MsgInitiateTokenDeposit } from "src/codegen/opinit/ophost/v1/tx";
-import { ClientState, SkipClientOptions } from "../state";
+import { ClientState, SkipClientOptions } from "../state/clientState";
 import { createRequestClient } from "../utils/generateApi";
-import { Affiliate, ChainAffiliates } from "../types/swaggerTypes";
-import { Fetch } from "src/utils/fetchClient";
+import { ApiState } from "src/state/apiState";
 
 export const setClientOptions = (options: SkipClientOptions = {}) => {
-  Fetch.client = createRequestClient({
+  ApiState.client = createRequestClient({
     baseUrl: options.apiUrl || "https://api.skip.build",
     apiKey: options.apiKey,
   });
@@ -41,46 +40,5 @@ export const setClientOptions = (options: SkipClientOptions = {}) => {
     ...(options.registryTypes ?? []),
   ]);
 
-  if (options.chainIdsToAffiliates) {
-    ClientState.cumulativeAffiliateFeeBPS = validateChainIdsToAffiliates(
-      options.chainIdsToAffiliates,
-    );
-    ClientState.chainIdsToAffiliates = options.chainIdsToAffiliates;
-  }
-
-  Fetch.setClientInitialized();
+  ApiState.setClientInitialized();
 };
-
-function validateChainIdsToAffiliates(chainIdsToAffiliates: Record<string, ChainAffiliates>) {
-  const affiliatesArray = Object.values(chainIdsToAffiliates)
-    .map((chain) => chain.affiliates)
-    .filter((a) => a !== undefined) as Affiliate[][];
-
-  const firstAffiliateBasisPointsFee = affiliatesArray[0]?.reduce((acc, affiliate) => {
-    if (!affiliate.basisPointsFee) {
-      throw new Error("basisPointFee must exist in each affiliate");
-    }
-    return acc + parseInt(affiliate.basisPointsFee, 10);
-  }, 0);
-
-  const allBasisPointsAreEqual = affiliatesArray.every((affiliate) => {
-    const totalBasisPointsFee = affiliate.reduce((acc, affiliate) => {
-      if (!affiliate.basisPointsFee) {
-        throw new Error("basisPointFee must exist in each affiliate");
-      }
-      if (!affiliate.address) {
-        throw new Error("address to receive fee must exist in each affiliate");
-      }
-      return acc + parseInt(affiliate?.basisPointsFee, 10);
-    }, 0);
-    return totalBasisPointsFee === firstAffiliateBasisPointsFee;
-  });
-
-  if (!allBasisPointsAreEqual) {
-    throw new Error(
-      "basisPointFee does not add up to the same number for each chain in chainIdsToAffiliates",
-    );
-  }
-
-  return firstAffiliateBasisPointsFee?.toFixed(0);
-}
