+++ packages/client/src/state/apiState.ts
@@ -1,22 +1,26 @@
-import { createRequestClient } from "./generateApi";
+import { ChainAffiliates } from "src/types/swaggerTypes";
+import { createRequestClient } from "../utils/generateApi";
 
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class
-export class Fetch {
+export class ApiState {
   static client: ReturnType<typeof createRequestClient>;
+
+  static chainIdsToAffiliates?: Record<string, ChainAffiliates>;
+  static cumulativeAffiliateFeeBPS?: string = "0";
 
   static isInitialized = false;
   static resolveInitialization: () => void;
   static clientInitialized: Promise<void> = new Promise<void>((resolve) => {
-    Fetch.resolveInitialization = () => {
-      if (!Fetch.isInitialized) {
-        Fetch.isInitialized = true;
+    ApiState.resolveInitialization = () => {
+      if (!ApiState.isInitialized) {
+        ApiState.isInitialized = true;
         resolve();
       }
     };
   });
 
   static setClientInitialized() {
-    Fetch.resolveInitialization();
+    ApiState.resolveInitialization();
   }
 }
